+++ changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
-**./test-parts/part-f-controller_models/models/m-model_definitions.js**
-* checkModelsLoop
-	* Merged 'currentQuote' variable into 'currentDesc'
+**./test-parts/part-g-controller_files/files/con-device_settings.js**
+* Renamed 'checkSignalValidationResult' function to 'handleSignalResult'
+
+---
+
+**./test-parts/part-g-controller_files/files/con-rio_index_node_list.js**
+* verifyNodeListCalled
+	* Removed `not.be.undefined` check on 'args[0]'
+
+---
+
+**./test-parts/part-g-controller_files/files/con-settings.js**
+* handlePathProperties
+	* Removed 'commonFunctionsFile.testObjectPropertyContent' calls.